Realised species models not accounting properly for covg due to errors!
Specifically when you have a mixture, the minor pop will get some coverage
due to errors on the major pop. At least one of the 3 false-calls for minor-pops
I was getting was for a sample with v high (300x) covg, and this highlighted the issue.
This commit gives a partial fix.