Various changes while debuggin - one major fix
The S model was using lambda_g not lambda_e when calculating the penalty.
This version of the code overcalls R considerably for genes, and r to a lesser extent.
There's also a bit of overcalling  R for mutations but let's not think about that for now
Fixing the gene models first!