Attempt to fix issues with species id that I previously introduced
I found a bug in t likelihood fo the Aureus model which was makoing it oo unlikely in a prevous commit. I then found that when mixing MRSE and MSSA, that the MRSE was never spotted. This commit fixes that fix. Tested on pure Staph epi and saprophyticus and an 80:20 mix of aureus and epi., plus one derivation set sample.
Not yet run on whole val/deriv sets