Fixed the bug that was causing false Non-MTBC calls.
With error cleaning off and InSiligoOligos we were observing that some samples that were previously being correctly called as MTB now were being called as NonMTBC.

The reason for this turned out to be that we had low median but high percentage coverage (100%) on e.g. lineage3 (presumably due to errors). This resulted in us choosing this lineage as our 'best hit' despite the fact that lineage 4 had equally high percentage covered and higher median coverage.
Since the median coverage on lineage3 was not >>> expected due to errors we call NonMTBC.

This was fixed by changing the logic for choosing the best hit. Changed > previous per_covg to >= previous_per_covg and updated when median covg was also > than previous hit.

In our example case we now correctly call lineage4 as our best hit and since it's median coverage >>> expected due to errors we call this as our lineage/species.

Please enter the commit message for your changes. Lines starting