Fix bug in parsing alternate R alleles
Something we thought we had fixed before: we were seeing an R allele
for which we have 100% perc covg, and median 1,
and then later on we were seeing another with 100% perc covg and median 84.
But we did not update our "best so far" variable with the second one,
because it did not have a striclty greater percentage covg.
However confusingly, we were still calling R, because the likelihood
function was going through the vob again and recalculating the best allele
median covg.
This is the sample Phelim spotted this with C00014844.
This now results in calling this sample being called R
with covg 8 on S and 84 on R allele. I think still an open issue as
to why not called r.