Lowered panel thresholds for nanopore.
Slightly changed the logic of get_best hit to prioties percentage covered.

There was an issue that arose where if there was two panels with the exact same median coverage but with differing percentage covered the first won even if the 2nd had a >> percentage covered.

Changed this to, update if percentage coverage is STRICTLY GREATER or if percentage coverage is >= and median coverage is >.