Change to underlying model.
Use weighted counts of SNPs to walk through tree.

Now use difference in frequency between in and outgroup rather than only looking at SNPs that NEVER appear in outgroup.
This should give more power to decide where, for example, there's a 99% freq in ingroup and 1% freq in outgroup.

However, this change means in a case where there's a SNP in 1 samples in each clade but clade 1 has 2 SNPs and clade 2 has 3 we would choose clade 1.