Models for resistotyping in the context of contamination
Some major changes bumped version up to 0.2.

When resistotyping at genes and mutations we now consider two coverages.

1) The expected coverage on target species (as before)

and

2) The maximum coverage on all non-target species (called contamination_covg)

Then if contamination_covg > 0 we compare 2 models for S taking the ML model:

Is the resistant coverage due to contamination with S from target

	    double llk_S_contaim = get_log_lik_R_S_coverage(var, contamination_covg, expected_covg, kmer);

Is the resistant coverage due to errors with S from target

    double llk_S_error = get_log_lik_R_S_coverage(var, expected_covg * err_rate / 3, expected_covg, kmer);

We then take the most likely of these for our llk_S model

And simillarly for R

Is the resistant coverage due to target with S from errors
    double llk_R_error = get_log_lik_R_S_coverage(var, expected_covg, expected_covg * err_rate / 3, kmer);
Is the resistant coverage due to target with S from contamination
    double llk_R_contaim = get_log_lik_R_S_coverage(var, expected_covg, contamination_covg, kmer);

The case with no contamination is unchaged.